+++ src/internal.h
@@ -39,11 +39,11 @@ _shiz_str_to_upper(char * string)
 }
 
 /**
- * @brief Determine whether two floats are approximately equal.
+ * @brief Determine whether two floating point values are approximately equal.
  */
 static inline
 bool
-_shiz_fequal(float const a, float const b)
+_shiz_fequal(double const a, double const b)
 {
     return (fabs(b - a) < SHIZEpsilon);
 }
